requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
pdf2image==1.16.3
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# AI/ML
openai==1.3.5
transformers==4.35.2
torch==2.1.0
sentence-transformers==2.2.2

# Utilities
numpy==1.24.3
requests==2.31.0
aiofiles==23.2.1
tenacity==8.2.3

# Frontend
streamlit==1.28.1
plotly==5.18.0